last

+++ Book_Delete.php
@@ -19,12 +19,20 @@ if (!$conn) {
             exit;
     }
 	
-	$result = mysqli_query($conn, "DELETE FROM Book WHERE ISBN = $ISBN");
+	$result = mysqli_query($conn, "SELECT ISBN FROM Book WHERE ISBN = $ISBN");
+	
+	if (mysqli_num_rows($result)==0) {
+		echo "The given ISBN does not refer to any book of the Library";
+	}
+	
+	else {
+	  $result = mysqli_query($conn, "DELETE FROM Book WHERE ISBN = $ISBN");
 	
 	if($result === FALSE) {
          die(mysqli_error($con)); 
     }
 	
 	echo "Book deleted";
+	}
 	
 	?>
